banat/requirements.txt
pandas
Requests
streamlit
google-generativeai
protobuf
